ibc_core_channel::types::proto
Module v1

Modules

acknowledgement

Nested message and enum types in Acknowledgement.

Structs

Acknowledgement

Acknowledgement is the recommended acknowledgement format to be used by app-specific protocols. NOTE: The field numbers 21 and 22 were explicitly chosen to avoid accidental conflicts with other protobuf message formats used for acknowledgements. The first byte of any message with this format will be the non-ASCII values 0xaa (result) or 0xb2 (error). Implemented as defined by ICS: https://github.com/cosmos/ibc/tree/master/spec/core/ics-004-channel-and-packet-semantics#acknowledgement-envelope

Channel

Channel defines pipeline for exactly-once packet delivery between specific modules on separate blockchains, which has at least one end capable of sending packets and one end capable of receiving packets.

Counterparty

Counterparty defines a channel end counterparty

ErrorReceipt

ErrorReceipt defines a type which encapsulates the upgrade sequence and error associated with the upgrade handshake failure. When a channel upgrade handshake is aborted both chains are expected to increment to the next sequence.

GenesisState

GenesisState defines the ibc channel submodule’s genesis state.

IdentifiedChannel

IdentifiedChannel defines a channel with additional port and channel identifier fields.

MsgAcknowledgement

MsgAcknowledgement receives incoming IBC acknowledgement

MsgAcknowledgementResponse

MsgAcknowledgementResponse defines the Msg/Acknowledgement response type.

MsgChannelCloseConfirm

MsgChannelCloseConfirm defines a msg sent by a Relayer to Chain B to acknowledge the change of channel state to CLOSED on Chain A.

MsgChannelCloseConfirmResponse

MsgChannelCloseConfirmResponse defines the Msg/ChannelCloseConfirm response type.

MsgChannelCloseInit

MsgChannelCloseInit defines a msg sent by a Relayer to Chain A to close a channel with Chain B.

MsgChannelCloseInitResponse

MsgChannelCloseInitResponse defines the Msg/ChannelCloseInit response type.

MsgChannelOpenAck

MsgChannelOpenAck defines a msg sent by a Relayer to Chain A to acknowledge the change of channel state to TRYOPEN on Chain B. WARNING: a channel upgrade MUST NOT initialize an upgrade for this channel in the same block as executing this message otherwise the counterparty will be incapable of opening.

MsgChannelOpenAckResponse

MsgChannelOpenAckResponse defines the Msg/ChannelOpenAck response type.

MsgChannelOpenConfirm

MsgChannelOpenConfirm defines a msg sent by a Relayer to Chain B to acknowledge the change of channel state to OPEN on Chain A.

MsgChannelOpenConfirmResponse

MsgChannelOpenConfirmResponse defines the Msg/ChannelOpenConfirm response type.

MsgChannelOpenInit

MsgChannelOpenInit defines an sdk.Msg to initialize a channel handshake. It is called by a relayer on Chain A.

MsgChannelOpenInitResponse

MsgChannelOpenInitResponse defines the Msg/ChannelOpenInit response type.

MsgChannelOpenTry

MsgChannelOpenInit defines a msg sent by a Relayer to try to open a channel on Chain B. The version field within the Channel field has been deprecated. Its value will be ignored by core IBC.

MsgChannelOpenTryResponse

MsgChannelOpenTryResponse defines the Msg/ChannelOpenTry response type.

MsgChannelUpgradeAck

MsgChannelUpgradeAck defines the request type for the ChannelUpgradeAck rpc

MsgChannelUpgradeAckResponse

MsgChannelUpgradeAckResponse defines MsgChannelUpgradeAck response type

MsgChannelUpgradeCancel

MsgChannelUpgradeCancel defines the request type for the ChannelUpgradeCancel rpc

MsgChannelUpgradeCancelResponse

MsgChannelUpgradeCancelResponse defines the MsgChannelUpgradeCancel response type

MsgChannelUpgradeConfirm

MsgChannelUpgradeConfirm defines the request type for the ChannelUpgradeConfirm rpc

MsgChannelUpgradeConfirmResponse

MsgChannelUpgradeConfirmResponse defines MsgChannelUpgradeConfirm response type

MsgChannelUpgradeInit

MsgChannelUpgradeInit defines the request type for the ChannelUpgradeInit rpc WARNING: Initializing a channel upgrade in the same block as opening the channel may result in the counterparty being incapable of opening.

MsgChannelUpgradeInitResponse

MsgChannelUpgradeInitResponse defines the MsgChannelUpgradeInit response type

MsgChannelUpgradeOpen

MsgChannelUpgradeOpen defines the request type for the ChannelUpgradeOpen rpc

MsgChannelUpgradeOpenResponse

MsgChannelUpgradeOpenResponse defines the MsgChannelUpgradeOpen response type

MsgChannelUpgradeTimeout

MsgChannelUpgradeTimeout defines the request type for the ChannelUpgradeTimeout rpc

MsgChannelUpgradeTimeoutResponse

MsgChannelUpgradeTimeoutRepsonse defines the MsgChannelUpgradeTimeout response type

MsgChannelUpgradeTry

MsgChannelUpgradeTry defines the request type for the ChannelUpgradeTry rpc

MsgChannelUpgradeTryResponse

MsgChannelUpgradeTryResponse defines the MsgChannelUpgradeTry response type

MsgPruneAcknowledgements

MsgPruneAcknowledgements defines the request type for the PruneAcknowledgements rpc.

MsgPruneAcknowledgementsResponse

MsgPruneAcknowledgementsResponse defines the response type for the PruneAcknowledgements rpc.

MsgRecvPacket

MsgRecvPacket receives incoming IBC packet

MsgRecvPacketResponse

MsgRecvPacketResponse defines the Msg/RecvPacket response type.

MsgTimeout

MsgTimeout receives timed-out packet

MsgTimeoutOnClose

MsgTimeoutOnClose timed-out packet upon counterparty channel closure.

MsgTimeoutOnCloseResponse

MsgTimeoutOnCloseResponse defines the Msg/TimeoutOnClose response type.

MsgTimeoutResponse

MsgTimeoutResponse defines the Msg/Timeout response type.

MsgUpdateParams

MsgUpdateParams is the MsgUpdateParams request type.

MsgUpdateParamsResponse

MsgUpdateParamsResponse defines the MsgUpdateParams response type.

Packet

Packet defines a type that carries data across different chains through IBC

PacketId

PacketId is an identifer for a unique Packet Source chains refer to packets by source port/channel Destination chains refer to packets by destination port/channel

PacketSequence

PacketSequence defines the genesis type necessary to retrieve and store next send and receive sequences.

PacketState

PacketState defines the generic type necessary to retrieve and store packet commitments, acknowledgements, and receipts. Caller is responsible for knowing the context necessary to interpret this state as a commitment, acknowledgement, or a receipt.

Params

Params defines the set of IBC channel parameters.

QueryChannelClientStateRequest

QueryChannelClientStateRequest is the request type for the Query/ClientState RPC method

QueryChannelClientStateResponse

QueryChannelClientStateResponse is the Response type for the Query/QueryChannelClientState RPC method

QueryChannelConsensusStateRequest

QueryChannelConsensusStateRequest is the request type for the Query/ConsensusState RPC method

QueryChannelConsensusStateResponse

QueryChannelClientStateResponse is the Response type for the Query/QueryChannelClientState RPC method

QueryChannelParamsRequest

QueryChannelParamsRequest is the request type for the Query/ChannelParams RPC method.

QueryChannelParamsResponse

QueryChannelParamsResponse is the response type for the Query/ChannelParams RPC method.

QueryChannelRequest

QueryChannelRequest is the request type for the Query/Channel RPC method

QueryChannelResponse

QueryChannelResponse is the response type for the Query/Channel RPC method. Besides the Channel end, it includes a proof and the height from which the proof was retrieved.

QueryChannelsRequest

QueryChannelsRequest is the request type for the Query/Channels RPC method

QueryChannelsResponse

QueryChannelsResponse is the response type for the Query/Channels RPC method.

QueryConnectionChannelsRequest

QueryConnectionChannelsRequest is the request type for the Query/QueryConnectionChannels RPC method

QueryConnectionChannelsResponse

QueryConnectionChannelsResponse is the Response type for the Query/QueryConnectionChannels RPC method

QueryNextSequenceReceiveRequest

QueryNextSequenceReceiveRequest is the request type for the Query/QueryNextSequenceReceiveRequest RPC method

QueryNextSequenceReceiveResponse

QuerySequenceResponse is the response type for the Query/QueryNextSequenceReceiveResponse RPC method

QueryNextSequenceSendRequest

QueryNextSequenceSendRequest is the request type for the Query/QueryNextSequenceSend RPC method

QueryNextSequenceSendResponse

QueryNextSequenceSendResponse is the request type for the Query/QueryNextSequenceSend RPC method

QueryPacketAcknowledgementRequest

QueryPacketAcknowledgementRequest is the request type for the Query/PacketAcknowledgement RPC method

QueryPacketAcknowledgementResponse

QueryPacketAcknowledgementResponse defines the client query response for a packet which also includes a proof and the height from which the proof was retrieved

QueryPacketAcknowledgementsRequest

QueryPacketAcknowledgementsRequest is the request type for the Query/QueryPacketCommitments RPC method

QueryPacketAcknowledgementsResponse

QueryPacketAcknowledgemetsResponse is the request type for the Query/QueryPacketAcknowledgements RPC method

QueryPacketCommitmentRequest

QueryPacketCommitmentRequest is the request type for the Query/PacketCommitment RPC method

QueryPacketCommitmentResponse

QueryPacketCommitmentResponse defines the client query response for a packet which also includes a proof and the height from which the proof was retrieved

QueryPacketCommitmentsRequest

QueryPacketCommitmentsRequest is the request type for the Query/QueryPacketCommitments RPC method

QueryPacketCommitmentsResponse

QueryPacketCommitmentsResponse is the request type for the Query/QueryPacketCommitments RPC method

QueryPacketReceiptRequest

QueryPacketReceiptRequest is the request type for the Query/PacketReceipt RPC method

QueryPacketReceiptResponse

QueryPacketReceiptResponse defines the client query response for a packet receipt which also includes a proof, and the height from which the proof was retrieved

QueryUnreceivedAcksRequest

QueryUnreceivedAcks is the request type for the Query/UnreceivedAcks RPC method

QueryUnreceivedAcksResponse

QueryUnreceivedAcksResponse is the response type for the Query/UnreceivedAcks RPC method

QueryUnreceivedPacketsRequest

QueryUnreceivedPacketsRequest is the request type for the Query/UnreceivedPackets RPC method

QueryUnreceivedPacketsResponse

QueryUnreceivedPacketsResponse is the response type for the Query/UnreceivedPacketCommitments RPC method

QueryUpgradeErrorRequest

QueryUpgradeErrorRequest is the request type for the Query/QueryUpgradeError RPC method

QueryUpgradeErrorResponse

QueryUpgradeErrorResponse is the response type for the Query/QueryUpgradeError RPC method

QueryUpgradeRequest

QueryUpgradeRequest is the request type for the QueryUpgradeRequest RPC method

QueryUpgradeResponse

QueryUpgradeResponse is the response type for the QueryUpgradeResponse RPC method

Timeout

Timeout defines an execution deadline structure for 04-channel handlers. This includes packet lifecycle handlers as well as the upgrade handshake handlers. A valid Timeout contains either one or both of a timestamp and block height (sequence).

Upgrade

Upgrade is a verifiable type which contains the relevant information for an attempted upgrade. It provides the proposed changes to the channel end, the timeout for this upgrade attempt and the next packet sequence which allows the counterparty to efficiently know the highest sequence it has received. The next sequence send is used for pruning and upgrading from unordered to ordered channels.

UpgradeFields

UpgradeFields are the fields in a channel end which may be changed during a channel upgrade.

Enums

Order

Order defines if a channel is ORDERED or UNORDERED

ResponseResultType

ResponseResultType defines the possible outcomes of the execution of a message

State

State defines if a channel is in one of the following states: CLOSED, INIT, TRYOPEN, OPEN, FLUSHING, FLUSHCOMPLETE or UNINITIALIZED.